Resolved all merging issues

+++ src/main/java/meal/planner/dataBase/itemList/ItemList.java
@@ -61,13 +61,13 @@ public class ItemList<T extends Item, S> {
 	}
 	
 	/** Returns the item with the given id */
-	public T findId(int id){
+	public T findId(long id){
 		assert isOk();
 		return findId(id, items);
 	}
 	
 	/** Returns the item with the given id */
-	private T findId(int id, ArrayList<T> items){
+	private T findId(long id, ArrayList<T> items){
 		assert isOk();
 		for(T t : items){
 			if(t.getId()==id){
